Add default value on RepositoryTupleField, do not need config if same value and field name (#31157)
