+++ features/encrypt/core/src/test/java/org/apache/shardingsphere/encrypt/it/EncryptConfigurationRepositoryTupleSwapperEngineIT.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.encrypt.it;
 
 import org.apache.shardingsphere.encrypt.yaml.config.YamlEncryptRuleConfiguration;
-import org.apache.shardingsphere.infra.util.yaml.datanode.RepositoryTuple;
+import org.apache.shardingsphere.mode.tuple.RepositoryTuple;
 import org.apache.shardingsphere.infra.yaml.config.pojo.rule.YamlRuleConfiguration;
 import org.apache.shardingsphere.test.it.yaml.RepositoryTupleSwapperEngineIT;
