+++ mode/api/src/main/java/org/apache/shardingsphere/mode/tuple/annotation/RepositoryTupleType.java
@@ -15,9 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.yaml.config.pojo.rule.annotation;
-
-import lombok.RequiredArgsConstructor;
+package org.apache.shardingsphere.mode.tuple.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -25,29 +23,16 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Repository tuple field.
+ * Repository tuple type.
  */
-@Target(ElementType.FIELD)
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface RepositoryTupleField {
+public @interface RepositoryTupleType {
     
     /**
      * Get persist path name.
-     * 
+     *
      * @return persist path name
      */
-    String value() default "";
-    
-    /**
-     * Get type.
-     * 
-     * @return type
-     */
-    Type type();
-    
-    @RequiredArgsConstructor
-    enum Type implements Comparable<Type> {
-        
-        ALGORITHM, DEFAULT_ALGORITHM, STRATEGY, DEFAULT_STRATEGY, DATA_SOURCE, TABLE, OTHER
-    }
+    String value();
 }
